Funktionen und Hilfsfunktionen ausgelagert. Programm getestet. In Helper sind Hilfsfunktionen, die aus der Gui aufgerufen werden müssen In Lib sind eigenständige Funktionen
