+++ GUI_IB_CNC_MPL.py
@@ -4,13 +4,15 @@ Created on Wed Jun 20 14:13:17 2018
 
 @author: halbauer
 """
+import os
 
 from PyQt5.QtWidgets import QApplication, QMainWindow, QFileDialog
 import sys  # We need sys so that we can pass argv to QApplication
 import numpy as np
-import os.path
 import axesMPL
 from Lib import IBtoGCode_Lib as IBtoGCode
+from Lib import IBtoGCode_Helper as IBtoGCode
+from Lib import dataObject as dOj
 
 
 class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
@@ -19,6 +21,9 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         self.setupUi(self)  # Laden der UI-Datei
         self.setup_triggers()  # lade Events für Buttons
         self.statusBar().showMessage('Programm erfolgreich geladen.')
+        self.df = None
+        self.var_save = None
+        self.data = dOj.DataFile()
 
     def setup_triggers(self):
         self.but_trim_csv.clicked.connect(self.trim_csv)
@@ -36,8 +41,8 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         self.but_out.clicked.connect(self.file_save)
 
     # Funktionen, die in ein extra Modul gehören
-    def test_if_string(self, path):
-        if IBtoGCode.test_if_string_helper(path):
+    def test_if_string(self):
+        if IBtoGCode.test_if_string_helper(self.data):
             self.statusBar().showMessage('Datei enthält ungültige Zeichen!')
             # self.status_line.setStyleSheet("color: rgb(181, 18, 62);")
             self.but_trim_csv.setEnabled(True)
@@ -47,13 +52,14 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
             self.but_init_data.setEnabled(True)
             self.but_QuickPlot.setEnabled(True)
 
-    def trim_csv(self, path, versuch, directory):
-        IBtoGCode.trim_csv_helper(path, versuch, directory)
-        self.proc_path(path)
-        self.test_if_string(path)
+    def trim_csv(self):
+        IBtoGCode.trim_csv_helper(self.data)
+        self.proc_path()
+        self.test_if_string()
 
-    def init_data(self, path):
-        df = IBtoGCode.init_data_helper(path, int(self.par_ang.text()))
+    def init_data(self):
+        df = IBtoGCode.init_data_helper(self.data.pfad, int(self.par_ang.text()))
+        self.data.set_data_frame(df)
 
         self.but_calc_ib.setEnabled(True)
         self.statusBar().showMessage('Datenmatrix erfolgreich erzeugt. Weiter mit CALC IB!', 2000)
@@ -64,7 +70,9 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         werte = [self.par_slp_out_1_1.text(), self.par_slp_out_1_2.text(), self.par_slp_out_1_3.text()]
         return IBtoGCode.out_to_np_helper(laengen, werte)
 
-    def calc_sq(self, df):
+    def calc_sq(self):
+        df = getattr(self.data, "df")
+
         # Übernehmen der Einträge aus der GUI
         i_0 = int(self.par_i0.text())
         foc_of = int(self.par_il.text())
@@ -74,6 +82,7 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         ang = int(self.par_ang.text())
         di = int(self.par_di.text())
         slp_out = self.out_to_np()
+        self.data.set_slp_out(slp_out)
 
         # Daten aus UI für die Funktion get_vs #################################
         par_vs = self.par_vs.text()
@@ -89,11 +98,12 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
 
         df = IBtoGCode.calc_sq_helper(df, i_0, foc_of, foc_ruhe, vs, slp_in, slp_out, ang, v_s_slope, di)
 
+        self.data.set_data_frame(df)
         self.but_filter_data.setEnabled(True)
         self.statusBar().showMessage('Strahlstromkurven berechnet. Weiter mit FILTER DATA!')
-        return df
 
-    def filter_data(self, df):
+    def filter_data(self):
+        df = getattr(self.data, "df")
         # --- Filterung der Daten ---
         df = IBtoGCode.filter_data_helper(self.filt_met_med.isChecked(), self.par_n_med.text(),
                                           self.filt_met_sav.isChecked(), self.par_n_sav.text(), self.par_p_sav.text(),
@@ -102,10 +112,11 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         self.statusBar().showMessage('Daten erfolgreich gefiltert. Weiter mit CALC CNC oder PLOT DATA!')
         return df
 
-    def create_cnc(self, df):
+    def create_cnc(self):
         ang = float(self.par_ang.text())
         d_ang = float(self.par_d_ang.text())
-        cnc = IBtoGCode.create_cnc_helper(ang, d_ang, df)
+        cnc = IBtoGCode.create_cnc_helper(ang, d_ang, getattr(self.data, "df"))
+        self.data.set_cnc(cnc)
         self.but_save_data.setEnabled(True)
         return cnc
 
@@ -113,11 +124,12 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
 
     def file_open(self):
         path = QFileDialog.getOpenFileName(self, 'Open File')[0]
+        self.data.set_pfad(path)
         if not path:
             return
-        # dir = os.path.dirname(path)
-        self.proc_path(path)
-        self.test_if_string(path)
+        self.data.set_directory(os.path.dirname(path))
+        self.proc_path()
+        self.test_if_string()
 
     def file_save(self):
         fname = f'{self.txt_versuch.text()}.MPF'
@@ -126,18 +138,17 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
             return
         self.txt_outfolder.setText(out_file)
 
-    def proc_path(self, path):
-        file, versuch = IBtoGCode.proc_path_helper(path)
-        self.txt_path.setText(path)
+    def proc_path(self):
+        file, versuch = IBtoGCode.proc_path_helper(self.data)
+        self.txt_path.setText(self.data.pfad)
         self.txt_file.setText(file)
         self.txt_versuch.setText(versuch)
 
     def quick_plot(self):
-        path = self.txt_path.text()
-        df = self.init_data(path)
-        df = self.calc_sq(df)
-        df = self.filter_data(df)
-        cnc = self.create_cnc(df)
+        self.init_data()
+        self.calc_sq()
+        df = self.filter_data()
+        cnc = self.create_cnc()
         self.plot_data(df, cnc)
         self.but_replot_data.setEnabled(True)
 
@@ -145,64 +156,26 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         cnc = self.cb_save_cnc.isChecked()
         csv = self.cb_save_csv.isChecked()
         if csv and cnc:
-            self.save_cnc()
-            self.save_csv()
+            self.save_cnc(self.data.versuch, self.data.directory, self.data.cnc)
+            IBtoGCode.save_csv(getattr(self.data, "df"), getattr(self.data, "versuch"), getattr(self.data, "directory"))
             self.statusBar().showMessage('CSV und CNC erfolgreich gesichert.', 2000)
         elif csv:
-            self.save_csv()
+            IBtoGCode.save_csv(getattr(self.data, "df"), getattr(self.data, "versuch"), getattr(self.data, "directory"))
             self.statusBar().showMessage('CSV erfolgreich gesichert.', 2000)
         elif cnc:
-            self.save_cnc()
+            self.save_cnc(getattr(self.data, "versuch"), getattr(self.data, "directory"), getattr(self.data, "cnc"))
             self.statusBar().showMessage('CNC-Code erfolgreich exportiert.', 2000)
         else:
             self.statusBar().showMessage('Keine Daten gesichert.', 2000)
 
-    def save_csv(self):
-        file_out = self.versuch + '_out.csv'
-        fname = os.path.join(self.dir + os.sep, file_out)
-        df_out = self.df[['Temperatur', 'Winkel', 'IB', 'IB-korr']].copy()
-        df_out.to_csv(path_or_buf=fname, sep=';', decimal=',',
-                      encoding='utf-8')
+    def save_cnc(self, versuch, direction, cnc):
+        pos_slope = int(IBtoGCode.get_slope_pos(getattr(self.data, "slp_out"))[0])
+        txt_out_folder = self.txt_outfolder.text()
+        ang = int(self.par_ang.text())
+        d_ang = float(self.par_d_ang.text())
+        slp_in = int(self.par_slp_in.text())
 
-    def save_cnc(self):
-        pos_slope = int(self.get_slope_pos()[0])
-        file_out = self.versuch + '_CNC.MPF'
-        if self.txt_outfolder.text() == '':
-            fname = os.path.join(self.dir, file_out)
-        else:
-            fname = os.path.join(self.txt_outfolder.text())
-
-        with open(fname, 'w') as f:  # direkte Erzeugung einer .MPF Datei
-            # f.write("IB_CURVE:\nG0 SL SLs)\n")    # Schreiben des Headers
-            f.write("G1 G91 G64 SL _SLo)\n")
-
-            cnc = self.cnc
-
-            for i in range(len(cnc)):
-                '''
-                zeilenweises Auslesen des Arrays und Erzeugung des CNC-Syntaxes
-                - 0: Spalte Winkel
-                - 2: Spalte Linsenstrom
-                - 3: Spalte Vorschubgeschwindigkeit
-                - 6: Spalte Strahlstrom
-                '''
-
-                ang = int(self.par_ang.text())
-                d_ang = float(self.par_d_ang.text())
-                slp_in = int(self.par_slp_in.text())
-
-                _A = f'A={d_ang}'
-                _SQ = f' SQ ({cnc["IB-korr2"].iloc[i]} + _SQ_off))'
-                _SL = f' SL {cnc["SL"].iloc[i]})'
-                _Fs = f' Fms {cnc["vs"].iloc[i]}'
-                if i == 0:
-                    f.write(_A + _SQ + _SL + _Fs + "\n")
-                elif cnc['Winkel'].iloc[i] <= slp_in:
-                    f.write(_A + _SQ + _SL + "\n")
-                elif cnc['Winkel'].iloc[i] >= (ang - pos_slope):
-                    f.write(_A + _SQ + _SL + _Fs + "\n")
-                else:
-                    f.write(_A + _SQ + _Fs + "\n")
+        IBtoGCode.save_cnc_helper(direction, versuch, txt_out_folder, d_ang, ang, slp_in, pos_slope, cnc)
 
     def enable_butt(self):
         self.par_n_med.setEnabled(False)
@@ -215,7 +188,7 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         self.par_p_sav.setEnabled(False)
 
     def save_var_state(self):
-        var_save = [
+        self.var_save = [
             self.par_i0.text(),
             self.par_di.text(),
             self.par_t_soll.text(),
@@ -238,7 +211,6 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
             self.par_n_sav.text(),
             self.par_p_sav.text()
         ]
-        return var_save
 
     def get_act_state(self):
         var_act = [
@@ -268,44 +240,45 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
 
     def replot(self):
         a = self.get_act_state()
-        b = self.save_var_state()
+        b = self.var_save
 
         if not np.array_equal(a[:14], b[:14]):
             self.calc_sq()
-            self.filter_data()
-            self.create_cnc()
-            self.replot_all()
+            df = self.filter_data()
+            cnc = self.create_cnc()
+            self.replot_all(df, cnc)
             print('Der Strahlstrom muss neu berechnet werden!')
         elif a[15] != b[15]:
             self.init_data()
             self.calc_sq()
-            self.filter_data()
-            self.create_cnc()
-            self.replot_all()
+            df = self.filter_data()
+            cnc = self.create_cnc()
+            self.replot_all(df, cnc)
             print('Alles muss neu berechnet werden!')
         elif not np.array_equal(a[16:], b[16:]):
-            self.filter_data()
-            self.create_cnc()
-            self.replot_all()
+            df = self.filter_data()
+            cnc = self.create_cnc()
+            self.replot_all(df, cnc)
             print('Die Ergebnisse müssen neu gefiltert werden!')
 
-    def replot_all(self):
+    def replot_all(self, df, cnc):
         w = self.graphicsView.canvas
         x_lim = self.ax1.get_xlim()
         y1_lim = self.ax1.get_ylim()
         y2_lim = self.ax2.get_ylim()
         w.figure.clf()
 
-        x = self.df['Winkel'].values
-        y1 = self.df['Temperatur'].values
-        y2_1 = self.df['IB'].values
-        y2_2 = self.df['IB-korr'].values
-        x2 = self.cnc['Winkel'].values
-        y2_3 = self.cnc['IB-korr2'].values
+        x = df['Winkel'].values
+        y1 = df['Temperatur'].values
+        y2_1 = df['IB'].values
+        y2_2 = df['IB-korr'].values
+        x2 = cnc['Winkel'].values
+        y2_3 = cnc['IB-korr2'].values
 
         ang = int(self.par_ang.text())
         t_soll = int(self.par_t_soll.text())
 
+        # todo ist die Definition von ax1 und ax2 an self in diesem Teil nötig?
         self.ax1 = w.figure.add_subplot(111)
         self.ax2 = self.ax1.twinx()
 
@@ -352,6 +325,7 @@ class MyApp(QMainWindow, axesMPL.Ui_MainWindow):
         ang = int(self.par_ang.text())
         t_soll = int(self.par_t_soll.text())
 
+        # todo ist die Definition von ax1 und ax2 an self in diesem Teil nötig?
         self.ax1 = w.figure.add_subplot(111)
         self.ax2 = self.ax1.twinx()
